Cleans up the orm hydrating and modifying stuff because it was hellish to understand. Also adds a default value for field, this closes #61 and closes #5.
For hydrating, there is now modify and populate, with accompanying _modify and _populate methods that can be easier overridden by child orms if needed. I removed get_modified(), hydrate(), and modify_fields() since they were no longer needed with a small bit of refactoring and it gets rid of a bunch of cruft that's just hung around for years.

For the default, you can now pass in a default=... to the Field instantiation and that will set the default value of that particular field, this involved a bit of refactoring in config.Field to support default values and adds an fdefault method and accompanying decorator to override it if you want but I wouldn't, the default fdefault method should do pretty much what everyone needs but I wanted it to match the rest of the interface